GCN/requirements.txt
absl-py @ file:///home/conda/feedstock_root/build_artifacts/absl-py_1623774798911/work
backcall==0.2.0
brotlipy==0.7.0
certifi==2021.5.30
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1606236101993/work
chardet @ file:///home/conda/feedstock_root/build_artifacts/chardet_1610093490430/work
cryptography @ file:///home/conda/feedstock_root/build_artifacts/cryptography_1616851476134/work
decorator==5.0.9
future @ file:///home/conda/feedstock_root/build_artifacts/future_1610147327521/work
googledrivedownloader @ file:///home/conda/feedstock_root/build_artifacts/googledrivedownloader_1619807768586/work
grpcio @ file:///home/conda/feedstock_root/build_artifacts/grpcio_1624380492467/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1593328102638/work
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1624799398687/work
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
jedi==0.18.0
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1621419064915/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1612898609989/work
jupyter-client==6.1.12
jupyter-core==4.7.1
Markdown @ file:///home/conda/feedstock_root/build_artifacts/markdown_1614595805172/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1621455677251/work
matplotlib-inline==0.1.2
networkx @ file:///home/conda/feedstock_root/build_artifacts/networkx_1598210780226/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1624399863733/work
olefile @ file:///home/conda/feedstock_root/build_artifacts/olefile_1602866521163/work
pandas==1.2.5
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1617383569452/work
protobuf==3.15.8
ptyprocess==0.7.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1593275161868/work
pygcn==0.1
Pygments==2.9.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1608055815057/work
pyparsing==2.4.7
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1610291447907/work
python-dateutil==2.8.1
python-louvain @ file:///home/conda/feedstock_root/build_artifacts/python-louvain_1609125227889/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1612179539967/work
PyYAML==5.4.1
pyzmq==22.1.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1608156231189/work
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1619628586492/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy_1624824888010/work
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
tensorboard==1.15.0
tensorboardX @ file:///home/conda/feedstock_root/build_artifacts/tensorboardx_1617715858833/work
threadpoolctl @ file:///tmp/tmp79xdzxkt/threadpoolctl-2.1.0-py3-none-any.whl
torch==1.8.1
torch-cluster==1.5.9
torch-geometric @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-geometric_1624734487033/work
torch-scatter==2.0.7
torch-sparse @ file:///usr/share/miniconda/envs/test/conda-bld/pytorch-sparse_1624691594596/work
torch-spline-conv==1.2.1
torchvision==0.9.1
tornado==6.1
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1623518955413/work
traitlets==5.0.5
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1624634538755/work
wcwidth==0.2.5
Werkzeug @ file:///home/conda/feedstock_root/build_artifacts/werkzeug_1621518206714/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1614945704755/work
